Symbian: don't merge native clipboard, overwrite.
If copying text from a Qt application that has both plaintext and
formatted text representations, then S60 applications only see the
plaintext.
If copying data from an S60 application, the Qt application can see
the plaintext. However as this was merged with the existing clipboard
contents in Qt, the formatted text was unchanged.
Now, we clear the internal clipboard QMimeData before writing the
contents of the symbian clipboard to it. This ensures stale data
from a previous copy is not left behind.

Task-Number: ou1cimx1#927246
Reviewed-By: mread