Symbian - prefer sessions started by this process to choose proxy
When WLAN and 3G connections are both active, the proxy for the wrong
connection may have been chosen in the case of plain sockets or
QNetworkAccessManager with an invalid configuration.
When enumarating active connections to choose a proxy, prefer a connection
that was opened by this process.

Task-number: QTBUG-22615
Task-number: ou1cimx1#930701
Reviewed-by: mread