tests: allow unstable tests to be marked with CONFIG+=insignificant_test
Marking a test with CONFIG+=insignificant_test will cause the exit code
of the test to be discarded during `make check'.  This is intended to be
used for tests which are valuable to run, but are known to be unstable
and are not feasible to immediately fix.

Change-Id: Id7b2407d0a13c8de19a58badb78fa7c018c9b50a
Reviewed-by: Jo Asplin
(cherry picked from commit a8469f14c26b6db70b97578655b22e4b3ec93cfc)
Reviewed-by: Jo Asplin <jo.asplin@nokia.com>
Reviewed-by: Pasi Pentikäinen <ext-pasi.a.pentikainen@nokia.com>