Reduce unnecessary QtCore DLL loads during Symbian app thread creation
The Symbain fast allocator integration in Qt 4.7 loads QtCore.DLL to
check availablility. This load has been causing a crash in the Nokia
Store app when a different version of QtCore 4.7 is installed. The new
DLL is loaded and it's static data tries to initialise before the
allocator is set up.

This change stores the allocator setup function so that extra DLL loads
are not required. Now the same allocator setup function is used for the
lifetime of the app.

Reviewed-by: Shane Kearns