Correct client rect calculation for MCL
The clientRect calculation was adjusting the rectangle for the status
pane. But this is now taken into account in the rectangle from Avkon,
so we do not need to adjust when the status pane is there.

However Avkon is supplying an inconsistent clientRect depending on
whether we have all Avkon panes disabled (status pane and CBAs).
When enabled, Avkon takes the splitview keyboard off the clientRect
space, when disabled, Avkon ignores the splitview keyboard. Since we
want to ignore the splitview keyboard effect on the clientRect, we
only adjust when the Avkon panes are enabled.

This has been tested with splitview keyboard, orientation change and
debug output of the client rectangle returned in the following
situations:

 - Normal app with status pane and CBA
 - App with CBA but invisible status pane
 - App with status pane and CBA disabled

Task-number: ou1cimx1#938780
Reviewed-by: Miikka Heikkinen