symbian - don't export qsymbianbearer.qtplugin twice
The file was being exported using both prj_exports and deployment
lines. As deployment is copying to epoc32 z and c directories, the
prj_exports is not needed.
Moved the prj_exports rules into the S60 5.0 section, where deployment
is not used but rather explicit rules added to the sis file.

Task-number: ou1cimx1#946574
Reviewed-by: mread