Symbian: allow apps to disable GL multisampling
On Symbian, QML apps had no way to disable GL multisampling if the
hardware supported it. This caused some apps to run out of graphics
memory and fail.

This change adds a way for apps to indicate that they don't want
multisampling, by setting an environment variable. For example:
  qputenv("QT_SYMBIAN_DISABLE_GL_MULTISAMPLE", "1");

This non-public API style was used as the use case is platform and
application specific, and not appropriate for public API.

Task-number: ou1cimx1#947771
Reviewed-by: Gareth Stockwell