Fix NB#290352 Qtwebprocess crashes @ QHttpNetworkReply::readAny
Several crash backtraces point to crash in QHttpNetworkReply::readAny,
where d->connection==0. This patch adds a check for d->connection to
QNetworkAccessHttpBackend. If the connection is found to be destroyed,
the request is finished.

Does not need to be merged to 4.8 because the internals have changed
(Peter Hartmann)

PMO 290352

Merge-request: 1491
Reviewed-by: Peter Hartmann <peter.hartmann@nokia.com>