Symbian: improving diagnostics for QWaitCondition
If a QWaitCondition was deleted while it was waiting, it would crash in
the wait() function while trying to use its internal mutex. This crash
is not particularly obvious in its cause.

QWaitCondition can detect this delete-while-waiting case, issue
appropriate warnings and avoid a crash in its internal state.

In the app crash case where this problem was found, there is a further
crash after this one due to an externally deleted mutex. The app is
clearly at fault. But at least there's now more debug info about
what the app is doing wrong.

Task-number: ou1cimx1#946509 partial fix
Reviewed-by: Shane Kearns