Symbian: Fix double deletion of selected text when using predictive
The selected text replacement was done incorrectly, leading to removal
of additional text equal to the lenght of the original selection.
Fixed by not doing any explicit replacement, but instead faking a
preedit character that is immediately removed, which triggers selected
text removal in the input method handling of the underlying edit
control.

Task-number: ou1cimx1#938665
Reviewed-by: Sami Merila