Symbian: Fix a case sensitivity issue with QMAKE_EXTRA_COMPILERS
The case of the drive letter generated to the absolute paths seems to
have changed between 4.7 and 4.8, which led into regression with
projects that utilize $$PWD in target name, as that still returns the
path with drive letter in same case as previously.
Fixed the critical check to be case insensitive in windows.

Task-number: ou1cimx1#951456
Reviewed-by: Murray Read