Default to QDir::homePath() in Symbian native file dialogs
Symbian places restrictions on the directories in which the native file
dialog can browse.  If the path passed to QFileDialog::getXxxFileName()
is inaccessible, the implementation defaults to a path which is
accessible.

Previously, this default path was QDir::rootPath().  Changes to the
Symbian file engine in Qt 4.8 mean that this now returns "C:\" which is
inaccesible to the file dialog.  This patch changes the default path to
QDir::homePath(), which returns "C:\data".

Task-number: ou1cimx1#947939
Reviewed-by: Miikka Heikkinen