Symbian - Fix QFile::map with non page aligned offsets
Although it is not documented, RFileMap native API requires the offset
into the file to be a multiple of the page size.
Round down the offset requested by the user to the page size.
Add back the difference before returning the base address to the user.

e.g. if the offset passed to QFile::map was 5000:
The offset passed to RFileMap will be 4096
The address returned to the user will be the RFileMap::Base() + 904

The modified address is used to key the hash of file mappings, so that
QFile::unmap() can be passed the modified address without needing changes.

Reviewed-by: mread
Task-number: ou1cimx#953054