Fixes QIconvCodec::convertToUnicode()
We need to hang onto the internal state of the utf16 converter
object whenever converting a long string to unicode. This was
failing with the text stream class -- the first read was okay,
but subsequent reads could not be correctly converted to unicode.

cherry-picked from qt5/qtbase: 72c07311228ce6eefb460d7e5da3de24220ba81e

Change-Id: I794d78808ccd9deedcb406bfe46e373c18142d05
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>