tests: allow unstable tests to be marked with CONFIG+=insignificant_test
Marking a test with CONFIG+=insignificant_test will cause the exit code
of the test to be discarded during `make check'. This is intended to be
used for tests which are valuable to run, but are known to be unstable
and are not feasible to immediately fix.

Task-number: QTQAINFRA-428
Change-Id: Idd45b04200182e3673195858ec08d31357dcdfaf
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
(cherry picked from commit 4c2d2b4cde3fef76056e528b87391f62acc8d928)
Reviewed-by: Pasi Pentikäinen <ext-pasi.a.pentikainen@nokia.com>