Recovering from draw failures after EGL surface creation failure
The Qt drawing system is not designed to handle failures in beginPaint
and endPaint. But the Symbian port is creating EGL contexts and surfaces
on demand in these functions. These can fail, eg in the defect reported
the app gets a visibile=true event for such a short period that it
fails to create the surface before it is invisible again. These failures
have to be coped with.

Brush image drawing is failing if the app is asked to draw
while in the background, because the EGL surface cannot be created.
After this, if the app comes to the foreground at the wrong time, the
failed drawing is shown on the screen as a flat colored area.

A failedToAlloc variable in the openvg paint engine was locking in this
failed drawing, so that it would never redraw. This is fixed by removing
all use of that variable.

Failed surface creation during a draw in response to a visibility
change event was also leaving the window backing store with bad content.
This is now cleared on a surface creation failure so that drawing
is retried when the next visibility change event arrives.

The retry mechanism for creating a surface on failure also
makes the problem circumstances more likely to occur and can block the
app unnecessarily. Now, instead of blocking for 1 second every time on
failed surface creation, it only blocks on the first failure in a
sequence thereby allowing other parts of the app to run freely while
it is in the background.

Task-number: ou1cimx1#951921
Change-Id: I1e27746957ff624b0d9a1cdc9b319d0a3477135d
Reviewed-by: Gareth Stockwell <gareth.stockwell@accenture.com>
Reviewed-by: Shane Kearns <ext-shane.2.kearns@nokia.com>
(cherry picked from commit 854e72538426b32f72acd6c6c76af04bc8b78994)
Reviewed-by: Juha Kukkonen
Reviewed-by: Pasi Pentikäinen <ext-pasi.a.pentikainen@nokia.com>