Update copyright year in translations.
It seems that in the past I have modified the source strings for
copyright messages, but have failed to also modify some of the
translated versions of those strings.  That resulted in some of the
strings being duplicated by Linguist and marked as obsolete. This time,
I have removed those obsolete duplicates and modified all of the
translated versions of the strings where the year is represented with
hindu-arabic numbers. The one exceptions is the farsi translation
(qt_fa.ts), for which I am unable to translate the year.

Change-Id: I224e9402f8cf1f02e301a6601a66680fd8fa5d0b
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@nokia.com>