Fix using custom paper sizes on Mac
This solves the bug with using custom paper sizes on Mac, if a QSize
was specified then with a Unit it would not respect this on Mac as it
would not convert it correctly.  Also paperSize() would return the
wrong value.

There is already a test for this in tst_QPrinter::testCustomPageSizes()

Task-number: QTBUG-18723
Change-Id: I47c0cf2dd48ea3e0daa3b748f337ec38714aaf34
Reviewed-by: Samuel Rødal <samuel.rodal@nokia.com>
Reviewed-by: Jiang Jiang <jiang.jiang@nokia.com>