directfb: Use QPlatformPixmap::fromFile for Qt resources
Move to fromFile to load Qt resources for two reasons. The
first is that ::fromData creates a QBuffer on our raw data
and is passing that to the QImageReaders. Right now we there
is a QFile -> QByteArray -> QBuffer transition that wastes
cycles. The other reason is that QPlatformPixmap::fromData
did/does not check if the decoded image is null and might
cause a crash because of that.

Change-Id: I800bfe4cd9b5d6a0d3531ed33f4228770703e94f
Original-Id: I7cb92d84874323e7205ab28883cc0fe9ecca1e27
Reviewed-by: Girish Ramakrishnan <girish.1.ramakrishnan@nokia.com>