Fix pinch-zoom problem when application is deactivated
Deactivation can happen for example if some notification dialog opens
on top of the application. In this case the PinchFinished signal was not
emitted.

It is fixed so that when the QDeclarativePinchArea component receives
QEvent::WindowDeactivate event, the ongoing touch activity is cancelled,
which causes the pinch to finish. Also if QEvent::TouchUpdate event is
received without receiving QEvent::TouchBegin event first, the
QEvent::TouchUpdate event will not cause the updatePinch() method to be
called.

Task-number: ou1cimx1#948030
Change-Id: Ib8b9e1a022415be06f1a8e778b887418f32c75ef
Reviewed-by: Murray Read <ext-murray.2.read@nokia.com>
Reviewed-by: Pasi Pentikäinen <ext-pasi.a.pentikainen@nokia.com>