Symbian - add rom upgrade flag to Qt sis files
The rom upgrade flag allows the sis file to be installed if there is
already a version of the dll in rom.
The flag is silently ignored unless the sis file is properly signed
with a certificate trusted for this purpose.
This change saves developers needing to edit pkg files after running
qmake.

Change-Id: I0914734c951ea1f0d1c48b169bee0b022bc9e275
Reviewed-by: Pasi Pentikäinen <ext-pasi.a.pentikainen@nokia.com>