Fix exception handling in Symbian event dispatcher
The Symbian event dispatcher handles exceptions from active objects
as well as leaves. But if the active object throws while it has
something pushed on the cleanup stack, the exception must be converted
to a leave for the cleanup to happen and the cleanup stack balance to
be maintained.

Task-number: ou1cimx1#970701
Change-Id: I6c8eed61cb1666e36eb4858a2c72ad0d8c6b0c8f
Reviewed-by: Shane Kearns <ext-shane.2.kearns@nokia.com>
(cherry picked from commit 8c5614dda55ed7b539e06a3846618df8255dd18c)
Reviewed-by: Pasi Pentikäinen <ext-pasi.a.pentikainen@nokia.com>