Change QUrl::toLocalFile to return path for relative urls
This change reverts the behaviour of QUrl::toLocalFile to the state
it was before 4.8. After this change the function returns the path if
the url is relative. Before this change an empty QString was returned.
A relative url can refer to a local file, but that can't be determined
from the url alone. Thus, it makes sense to return the path for such
urls.

Added documentation to explain that the function works like this to
maintain backward compatability in 4.x, but the handling of relative
URLs will change in 5.0.

Task-number: QTBUG-19827
Change-Id: I8bb8f4603a5936c0359afc1b6ff98824fad6cbc9
Reviewed-by: Olivier Goffart <ogoffart@woboq.com>
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>