Fixed QEventDispatcherQPA::processEvents()
QEventDispatcherQPA::processEvents() calls
QEventDispatcherUNIX::processEvents() twice in a row. Before the first call,
it processes gui, posted events, input events and so on. It then calls
QEventDispatcherUNIX::processEvents(), which will eventually drain the
thread_pipe. If events are added by a second thread to the event queue, right
after select() returns but before the thread_pipe is drained, once the program
enters the second call to QEventDispatcherUNIX::processEvent(), select() will
block, because the thread_pipe is empty, but the events will be stalled on the
queue until something else causes select to return, because they are not
handled between these two calls.

Change-Id: Ie9a74573559feff6480b630abc237a8e5c89484b
Reviewed-by: Robin Burchell <robin+qt@viroteck.net>
Reviewed-by: Sean Harmer <sh@theharmers.co.uk>
Reviewed-by: Bradley T. Hughes <bradley.hughes@nokia.com>