Fix progress bar values >100% when http upload is retried
If a http upload (PUT or POST) fails, the IO device is reset
and the upload is retried a second time.
Now we also reset the progress counter in the wrapper class
QNonContigiousByteDevice.
That way the progress bar goes back to 0% and stops at 100%
Previously if the 1st attempt failed at 20%, the progress values
would carry on from 20% and finish at 120%.

Task-number: ou1cimx1#970373
Task-number: QTBUG-23993
Change-Id: Ifc9963595113cf3001ed45f94f525e2305ad479b
Reviewed-by: Markus Goetz <markus@woboq.com>
Reviewed-by: Richard J. Moore <rich@kde.org>