Take into account shaping in findRealWindow
It can happen that there is a window covering all the screen but it is shaped
to only take part of the screen. If that happens, besides the condition of
   QRect(attr.x,attr.y,attr.width,attr.height).contains(pos)
we also need to query the server for its region rectangles and make
sure the cursor is inside one of those rectangles. If that does not happen
we have to return 0 so the hierarchical XQueryTree search continues

Change-Id: Icb2204a50197e4a5e02e75601c67287525b290b0
Reviewed-by: Lars Knoll <lars.knoll@nokia.com>