Fix limitting shortcuts to keypad only
The Qt::KeypadModifier modifier is internally masked away from all
shortcuts. So it is not possible to set a keypad only shortcut.
Changed the implementation so that first a full keysequence match is
searched. Then if no match is found the same sequence is tried
without the keypad modifer.

Added a autotest for this also to cover the basic use cases relating
to this.

This is a backport of qt5 commit:
547a1bea492954d828aa0798be93384669812489

Task-number: QTBUG-20191
Change-Id: Ibb6809050ff6da23bdb1c087b93e517d1963d34c
Reviewed-by: Stephen Kelly <stephen.kelly@kdab.com>