Fix QListView::scrollTo() when there are hidden rows
QListView does not consider hidden rows when scrolling to an item.
If there are hidden rows (or columns) before the selected item then
the visual index of an item is not the same as the row index
from the model. So scrolling will be off by the number of hidden
rows before the selected item.

Added a autotest for this also.

Task-number: QTBUG-21115
Change-Id: I49c39033d83f80d4405b59510617096029ee6d3b
Reviewed-by: Stephen Kelly <stephen.kelly@kdab.com>