Ensure the scrollers appears when there is too many menu items
When there are too many items in a single column menu then scrollers
will appear for the menu to enable scrolling to the other items.
However, this broke in Qt 4.8 as the scrollers would not appear anymore
due to the adjustment of the size of the menu to account for the size
of the desktop.

Task-number: QTBUG-23507
Change-Id: Ib6bb7a5572bc0fbbf3de6dac03a08b69404804a3
Reviewed-by: Pierre Rossi <pierre.rossi@nokia.com>