Fixes QProcess on QNX
This patch first re-enables childStartedPipe on QNX and uses, just like other
platforms, to monitor the inferior process life cycle. This also aligns
QNX-specific code logic closer to the common QProcess logic, making it
unecessary to have custom versions of functions such as
QProcess::waitForStarted(). The only difference that remains is the use of
spawn() instead of fork(), because both fork() and vfork() do not support
multi-threaded applications on QNX and will segfault.

Change-Id: I8e1f9823629fcb5e7c3c128fafc2542a403b969e
Reviewed-by: Sean Harmer <sh@theharmers.co.uk>
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>