Make configure.exe only detect each compiler once
When reading the registry on a 64 bit Windows, the same compiler
can be detected twice, breaking the -platform detection even when
only one compiler is in the path.
Fix this by taking advantage of the CompilerInfo struct ordering
and ignore detection of the same compiler.

Change-Id: I583230520d2e0859196f9d7c8af31adbb981a6ca
Reviewed-by: Joerg Bornemann <joerg.bornemann@nokia.com>
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@nokia.com>
(cherry picked from commit fa5cf01e7468e53508b15726625fd681892119a4)
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@digia.com>