Correcting fixed native orientation translation for touch points
For a symbian application using fixed native orientation held in
landscape, touches starting from the top of the screen were ignored.

This was caused by an off-by-one error when reflecting touch positions.
For example a touch in the range 0..359 could be mapped to the range
360..1. And since 0 mapped to a value outside of the valid range,
touches starting with that were ignored.

The same translation function is used for both touch points and
rectangle coordinates, but these use different meanings for a point.
For a touch, a point is the centre of the touched pixel, so translations
should be against the last pixel, which is at size-1. But Symbian TRect
points specify the edges of pixels, and these should be translated
against the edge of the screen, which is at size. So the function now
takes a parameter describing the type of translation required.

Task-number: ou1cimx1#982747
Change-Id: I7af10b3a5b9b8658db9af8fa9bc04ad9cb531588
Reviewed-by: Laszlo Agocs <laszlo.p.agocs@nokia.com>
Reviewed-by: Gareth Stockwell <ext-gareth.stockwell@nokia.com>