Enable storage of global Qt settings in app-local settings file on Mac
The Mac App Store has limitations on where applications can write their
settings, and com.trolltech.plist is not allowed. Changed the settings
code to store all settings in the app-local file when the application
runs in sandbox, or when the application's Info.plist contains a key
"ForAppStore" with value "yes". The application's bundle identifier is
also used for naming the settings file in these cases.

Task-number: QTBUG-16549
Change-Id: Idd2241fbd7eb346da987226f05460642b0d6e5a3
Reviewed-by: Morten Johan Sørvig <morten.sorvig@nokia.com>