Removing Symbian fast allocator implementation
An implementation of the Symbian fast allocator was added to Qt
before it was available in all Symbian devices, to allow Qt
the benefit of it in all installations.

Now it is available in all Symbian devices and there is no
need to keep a copy in Qt. So it is being removed to simplify
the code.

Task-number: ou1cimx1#946699
Change-Id: I937faca61beb4ef66a858766ebfa3deff564c786
Reviewed-by: Shane Kearns <shane.kearns@accenture.com>