Don't allow app panic on QProcess destruction, kill or terminate
QProcess destruction will call kill() if the process is still running.
PRocess::Kill() and Terminate() both require the PowerMgmt capability
to operate on Symbian, otherwise a KERN-EXEC 46 panic happens.

An app should be able to use QProcess safely if it doesn't have
PowerMgmt capability, it should just be prevented from trying to kill
the created process. Now a debug message is issued on ~QProcess(),
kill() or terminate() if the capability is not present, rather than
actually trying to kill/terminate the process.

Task-number: ou1cimx1#985227
Change-Id: I3242ca2c39528c70c2c79e39f6a6384dd72f6ae6
Reviewed-by: Shane Kearns <shane.kearns@accenture.com>