BlackBerry Plugin: support 8 mouse buttons, instead of just 3.
The mask of possible mouse buttons in QNX provides tracking for the
up/down State of up to 8 mouse buttons. This update adds support
or 2 of buttons which we previously ignored in Qt on this Platform.

Backport of a0933f4d7485d22f38b80c67f79b8d3f721b19a2 in Qt5/qtbase
reduced to the 5 buttons Qt4 has support for

Related to task-number: QTBUG-24682

Change-Id: I4253b96b2d7ed3f806749432b2eb7f884fd0f580
Reviewed-by: Robin Burchell <robin+qt@viroteck.net>
Reviewed-by: Rick Stockton <rickstockton@reno-computerhelp.com>
Reviewed-by: Rafael Roquetto <rafael.roquetto@kdab.com>