Cherry picking qt5 bugfix 437da962c2b486b1a1722aac003bfa00b93ac204 to qt4.
Original description for bugfix:
Fix bug showing folder icons for files

Fix an issue where the cache was updated with the folder icon,
even though this is never used from the cache. This caused
files to sometimes be shown with a folder icon: If a folder
has an extension (e.g. a folder named 'folder.ext'), this
means the folder icon will be written to the cache for the
'ext' extension, and from that point on all .ext files will
be shown with the folder icon.

The fix is to not save an icon in the cache for those cases
where it will not be used from the cache anyway, by using the
same condition for updating the cache as for reading it (which
is in line 8 of the same function).

Change-Id: I349526ff017b1a2409e3dcc494b86bb541740790
Reviewed-by: Stephen Kelly <stephen.kelly@kdab.com>