Give QThread threads better names on Symbian
QThread sets the name of a Symbian thread that it creates. This name
can appear in a number of debugging scenarios, and it can give useful
information. However the existing name set by QThread has not been
informative.

This change improves the amount of information in the thread name.
Threads are now named as:
    <object name>_<class name>_v=<class vptr>_<random>
The class name and vptr can be used to identify the QThread subclass
type. In the case of a QThread subclass that generates a .moc file,
the class name will be the true class name. However for other cases,
eg where QThread is used directly, or is subclassed without a QOBJECT
declaration, the name may simply be "QThread", and the vptr field
might then be more informative.

This change also allows the use of RTTI to get the true class name,
through a rebuild of Qt, defining the macro name
QT_USE_RTTI_IN_THREAD_CLASSNAME. This is not enabled by default, as
there may be Symbian cases where RTTI does not work.

Task-number: QTBUG-24950
Change-Id: Ifae9c6cb64638e95a01d34e421aa36f6fd0e7444
Reviewed-by: Shane Kearns <shane.kearns@accenture.com>
Reviewed-by: Gareth Stockwell <ext-gareth.stockwell@nokia.com>