Fix window focusing problem in Symbian
The application window didn't receive focus properly on Symbian when
minimized window was maximized. This was because the code first tried
to set the focus and then made the window visible. However, invisible
window doesn't accept focus, so the visibility should be set first
and focus only after that.

Not having focus properly set caused software input pane to fail to
start in the Camera application, as was reported in the error case
ou1cimx1#989756.

Change-Id: Id659662e7d92205321746ed5a226843382a6ac89
Reviewed-by: Murray Read <ext-murray.2.read@nokia.com>
Reviewed-by: Juha Kukkonen <ext-juha.kukkonen@nokia.com>
Reviewed-by: Pasi Pentikäinen <ext-pasi.a.pentikainen@nokia.com>