Reduce virtual address space use by JavaScriptCore on Symbian
JavaScriptCore is reserving 128MB address space for each instance
of JSGlobalData that is created in an app. Eg there is one per
QDeclarativeView, via QDeclarativeEngine, via QScriptEngine.
This can contribute to the app running out
of address space and then crashing.

The AlignedBlockAllocator is modified to be a process singleton so that
all JSGlobalData objects share the same instance.
Now there is only one 128MB address space reservation.

Task-number: ou1cimx1#993804
Change-Id: I9e8702810541905d5e9be197a5f6a9fe8e9bd0c5
Reviewed-by: Shane Kearns <shane.kearns@accenture.com>
Reviewed-by: Pasi Pentikäinen <ext-pasi.a.pentikainen@nokia.com>